PROD-2117 Generate UUIDv4 on the client side for served notices and privacy preferences saved. Send served notices request async without waiting for response.
