+++ clients/fides-js/src/components/notices/NoticeOverlay.tsx
@@ -1,6 +1,7 @@
 import "../fides.css";
 
 import { FunctionComponent, h } from "preact";
+
 import { useCallback, useEffect, useMemo, useState } from "preact/hooks";
 
 import { getConsentContext } from "../../lib/consent-context";
@@ -21,7 +22,6 @@ import {
   updateCookieFromNoticePreferences,
 } from "../../lib/cookie";
 import { dispatchFidesEvent } from "../../lib/events";
-import { useConsentServed } from "../../lib/hooks";
 import {
   selectBestExperienceConfigTranslation,
   selectBestNoticeTranslation,
@@ -34,6 +34,7 @@ import Overlay from "../Overlay";
 import { OverlayProps } from "../types";
 import { NoticeToggleProps, NoticeToggles } from "./NoticeToggles";
 import { useI18n } from "../../lib/i18n/i18n-context";
+import { useConsentServed, useUUID4 } from "~/lib/hooks";
 
 /**
  * Define a special PrivacyNoticeItem, where we've narrowed the list of
@@ -148,7 +149,9 @@ const NoticeOverlay: FunctionComponent<OverlayProps> = ({
     };
   });
 
-  const { servedNotice } = useConsentServed({
+  const servedNoticeHistoryId = useUUID4();
+  useConsentServed({
+    servedNoticeHistoryId,
     privacyExperienceConfigHistoryId,
     privacyNoticeHistoryIds: privacyNoticeItems.reduce((ids, e) => {
       const id = e.bestTranslation?.privacy_notice_history_id;
@@ -197,7 +200,7 @@ const NoticeOverlay: FunctionComponent<OverlayProps> = ({
         options,
         userLocationString: fidesRegionString,
         cookie,
-        servedNoticeHistoryId: servedNotice?.served_notice_history_id,
+        servedNoticeHistoryId,
         updateCookie: (oldCookie) =>
           updateCookieFromNoticePreferences(
             oldCookie,
@@ -214,7 +217,7 @@ const NoticeOverlay: FunctionComponent<OverlayProps> = ({
       options,
       privacyExperienceConfigHistoryId,
       privacyNoticeItems,
-      servedNotice,
+      servedNoticeHistoryId,
     ]
   );
