+++ clients/fides-js/src/lib/common-utils.ts
@@ -1,3 +1,16 @@
 export const raise = (message: string) => {
   throw new Error(message);
 };
+
+/**
+ * Extracts the id value of each object in the list and returns a list
+ * of IDs, either strings or numbers based on the IDs' type.
+ */
+export const extractIds = <T extends { id: string | number }[]>(
+  modelList?: T
+): any[] => {
+  if (!modelList) {
+    return [];
+  }
+  return modelList.map((model) => model.id);
+};
