+++ clients/fides-js/src/lib/hooks/index.ts
@@ -1,0 +1,4 @@
+export { default as useConsentServed } from "./useConsentServed";
+export { default as useHasMounted } from "./useHasMounted";
+export { default as useUUID4 } from "./useUUID4";
+export { default as useDisclosure } from "./useDisclosure";
