+++ clients/fides-js/src/lib/hooks/useDisclosure.ts
@@ -1,0 +1,41 @@
+import { useCallback, useState } from "preact/hooks";
+
+/**
+ * Hook to facilitate showing/hiding while adhering to WAI
+ * based on chakra-ui's `useDisclosure`
+ */
+const useDisclosure = ({ id }: { id: string }) => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const onClose = useCallback(() => setIsOpen(false), []);
+  const onOpen = useCallback(() => setIsOpen(true), []);
+
+  const onToggle = useCallback(() => {
+    if (isOpen) {
+      onClose();
+    } else {
+      onOpen();
+    }
+  }, [isOpen, onOpen, onClose]);
+
+  const getButtonProps = () => ({
+    "aria-expanded": isOpen,
+    "aria-controls": id,
+    onClick: onToggle,
+  });
+
+  const getDisclosureProps = () => ({
+    id,
+    className: isOpen ? "fides-disclosure-visible" : "fides-disclosure-hidden",
+  });
+
+  return {
+    isOpen,
+    onOpen,
+    onClose,
+    onToggle,
+    getButtonProps,
+    getDisclosureProps,
+  };
+};
+export default useDisclosure;
