+++ clients/fides-js/src/lib/hooks/useHasMounted.ts
@@ -1,0 +1,18 @@
+import { useEffect, useState } from "preact/hooks";
+
+/**
+ * Hook which tracks if the app has mounted yet.
+ *
+ * Used to make sure the server and client UIs match for hydration
+ * Adapted from https://www.joshwcomeau.com/react/the-perils-of-rehydration/
+ */
+const useHasMounted = () => {
+  const [hasMounted, setHasMounted] = useState(false);
+
+  useEffect(() => {
+    setHasMounted(true);
+  }, []);
+
+  return hasMounted;
+};
+export default useHasMounted;
